core: adapt WatchHistory watch progress tracking
Track watch progress of each episode in percentages instead of a duration
according to the latest commit in ui.
Use a dedicated WatchProgress class to simplify possible future changes
to the way an episode watch progress is being tracked: if the actual
progress value type will change - only the WatchProgress class
will have to change.