core: improve the crutch that handles AllShowsEvent data integrity
Since shows are getting updated from a lot of different places - simply
always call the ViewAllShowsTask when the AllShowsEvent has to be sent
to the UserInterface (when the "back" action is performed and
a AllShowsEvent becomes a new top event).
Get rid of the previous version of this crutch, called "RePopulateAllShows".