core: restore text input field values after returning back to their views
Change ImportShowFromJsonEvent.getRawShowCrawler to
ImportShowFromJsonEvent.getShowCrawler to match the attribute name, expected
by the UI.
Make PreviewCrawlerTask update current EditCrawlerEvent with the crawler body
in order to display that body in the view, when the user returns back to it.
Make SaveShowCrawlerTask update current EditShowCrawlerEvent with the
show's name in order to display that show name in the view, when the
application fails to crawl the show and goes back to it.
Make ImportShowCrawlerTask update current ImportShowFromJsonEvent with
the show crawler body in order to display that body in the view, when
the application fails to crawl the show and goes back to it.