ui: fixes to make desktop version work
Use char codes instead of literals for material icon names.
Make ShowCard be visible in the AllShows view.
Make ShowDetails title not disappear.
Remove scroll bar in the ShowPlayer.
Fix "startProgress" processing in ShowPlayer: wait for the first buffering
to finish and only after that perform the seek, otherwise the seek
will be ignored.