core: fix episode start progress calculation
In case the user plays the show for the first time - play each episode
from the beginning.
In case the user plays the show, he has already watched - play the
last watched episode from where the user has left it. Play all subsequent
episodes from the beginning.
In case the user replays the show from the beginning - play all the episodes
of the show from the beginning.
In case the user plays a specific episode of the show - play all the episodes
from the beginning.