Improve module naming in serio-qt
Make module naming as strict as it is in the core:
each sub-folder in serio-qt/src represent a module
and each of such module folders should have a class
with the same or at least almost the same name
as a directory it belongs to.
Rename network folder to http-client.
Rename concurrency folder to task-executor and rename
TaskExecutor class to QTaskExecutor to avoid name
collision with the future serio::core class.
Rename storage folder to database-storage.