Introduce possibility of displaying an icon in ListTile.qml
If TileModel or a TileModel-like object has it's attribute
"icon" set to an empty string - ListTile.qml will behave
like before this change: only title and subtitle will
get displayed.
If TileModel has it's "icon" set to "placeholder" -
an empty space will be displayed in place of the icon
in the ListTile.qml (useful when you have a list
of tiles where only some of the tiles have icons: it
preserves a common padding).
If TileModel has it's "icon" set to any of the material icon
names - corresponding material icon will be displayed in
ListTile.qml.
For now TileModel's icon can't be set, since nobody
needs it.