Get rid of image size scaling
Size scale works both ways: on high-resolution screens
it up-scales image to higher sizes, thus increases image
loading time (which is noticeable on 4K screen in TvShowView).
Get rid of the scaling (it seems to work okay even in cases
when the image is bigger than it has to be).