Fix incorrect StackOfViews handling on crawling error
When user attempts to create a crawler, that overrides
an existing tv show, if an error occurs during the crawl,
after dismissing an error view dialog user gets presented
with the tv show override dialog. This is incorrect: user
should be presented with either a TvShowCrawlerEditorView
or an ImportTvShowCrawlerView depending on how the user
has created the executed crawler (by constructing it
manually or by importing it).