Enable playing tv show from the last watched episode in tv-show-player
Continue playing last watched episode of the specified tv show from the
position, where the user has left it previously.
If user has completely watched last episode - play the episode next to it
from the beginning.
If user has not watched the specified tv show - play the tv show from the
first episode from the beginning.
If user selects specific episode to play - always play it from the beginning.
Make Player define the start offset in percentage of where the currently
played episode should start playing in the total duration of the episode.
Consider WatchProgress complete if it's percentage value is 100.