Create include directory of the core project
In the future some of the modules will be re-used in
some auxiliary executables (crawler runtime will have
it's own CLI executable to debug crawlers). To enable
that - create a proper include directory, that can
be used by those executables.
All the new module headers should be in that directory.