Refactor crawler runtime a little bit
Replace do while loop with something more common (like a normal while
loop).
Keep finished execution pointers typed (there is no reason not to do
so).
Re-structure the main loop in such a way, that there is no need
for CrawlerExecution to check if it has failed or not in every
executeStep() call.