Implement support for platforms with search not available to user
There can be platforms, that do not provide categories or search as a part
of their config (because they have low-quality tv-show thumbnails for
example). Such platforms only provide tv-show episode information.
For such platforms the 'available-to-user' option of the 'search' field
should be disabled, so that their search results with their low-quality
thumbnails do not appear in results when user searches for a tv-show.