Create SuggestionsCrawlerConfig
Introduce a separate crawler config type (formerly available as
"suggestionsCrawler" under the "tv-show" crawler config) -
SuggestionsCrawlerConfig, each platform can provide, that enables crawling
for tv-show's, that can be suggested to the user, watching another
tv-show.
Such crawlers will take the name of the currently watched tv-show as
an argument and will crawl list of tv-shows, related to the specified one.