Better handle plural localizations
Simplify Localization constructor API.
Introduce a mechanism of plurality-based parameter localization:
each specified translation can have an optional rule, that will
be applied to a parameter, with which the translation will be
formatted. In case a text is being formatted using a parameter -
find the first translation, rule of which matches the ending of
the parameter. Translations are search in the order in which they
have been specified:
in a list:
"translation 1", "1"
"translation 2", "11"
when formatting the text with parameter "11" - the first translation
will be picked regardless of the fact that the second translation
matches better.