Make localization store the current language as config property
Instead of storing current language in memory as a part of
Localization - store it in the Config, so that when
the application restarts, user-made changes to the selected
language would persist.
Make Localization instantiation fail in case no
translations have been provided - in such case it won't be
able to protect it's invariant and always return current
language (it needs to translations to determine it
in some cases).