Create a reusable FocusableList component
Create a reusable List component, that properly displays
the currently focused item. The FocusableList also
allows pre-selecting an item in the list, so that when
the list is rendered - the specified item is focused
by default.
Use the new list in Settings.js and in SelectDialog.js.
Get rid of the old ad-hoc implementations.