Fix flicker when switching from Settings to SelectDialog
Make ChildAppBar in Settings look the same it is in
SelectDialog.
Don't apply top margin to the content of the Settings
view: now ChildAppBar's background is invisible so
there is no visual clash of it with the Paper in the view,
thus there is no reason in distancing them further apart
anymore.