Partially implement the Search view
Allow specifying child elements in ChildAppBar.js.
Implement Search view, that for now only has a SearchBar.
Make CoreStub.js log every event received for debugging
purposes.
Make Settings and Search use different kinds of back event
types for now, so that when we press back in Search
we are not getting redirected to the Settings view.