Make execute_crawlers() slightly more readable
Rename start_profiling() to __profile_trace(), so that
it stands out among the other actual functions that
do stuff.
Re-order some functions, so that functions, that belong
to the same system, are executed together. This reduces
amount of __profile_trace() calls needed.
Change __profile_trace() signature:
- make system_id - the first argument, so that
in places, that call __profile_trace() it is easier
to read what system is being executed next
- make a whole crawler_runtime - the second argument
to reduce amount of code in places, that call
__profile_trace()