Animate title screen appearance
Also create an installDebugAndRun run script for cases
when you don't need Android Studio capabilities
(like debug, profiling or android-specific
auto-completion) and just need to run a new version
of an app and see if it works. In such cases it is
much faster then the studio.
The script builds debug version of the apk, installs
it on the device, currently connected to adb,
and runs it. The script only works when there is only
one device connected.
And exclude a bunch of folders from git in ui-web
project.