Implement EditTextDialog view
Make simple version, that just displays stuff and
does not do anything else. It is already able to
render externally specified content though.
Implement a way to request focus to a specific html
element when the time is right. Doing it inside the
create*() function is not such place, because
all kinds of stuff can happen during it's call and
actual displayElement invocation, like animations.
Hell, the element is not even attached to DOM
at that point, so focus() call will just get ignored.