Implement a separate Dialog view
Extract a Dialog from EditTextDialog and make it
a standalone view.
Allow Dialog having externally specifiable elements
inside of it between the description and the buttons.
Make buttons in dialog optional: if button's text
is not specified - the button will not be displayed.
Always focus the first button.
Make EditTextDialog use the new Dialog internally.