Implement view rendering mechanism on UI side
Create window-level displayView() function, that can
be called by the runtime, running the browser with
the UI.
Add "view" field to content, to let ui know which
view to render. Make all the functions, that request
ui to display a new view, only request ui to change
current view in case the specified content requests
their view to get rendered.
Make createDialog() still request ui to render dialog
view even if current content.view is not DIALOG in
case innerElements is specified (in such case
dialog view is not used as a standalone view but as
a part of some other view).