Fix view transition interruption bug
- A is transitioning out
- B is set to replace it
- B has ALL animations disabled
- request arrives to render C
Make animateElement() store transition animation of B in
nextViewAnimation instead of currentViewAnimation (while
A is still transitioning out). This way when C arrives
we won't attempt to immediately skip A's transition and
replace it with C (which will then make B replace C, since
B's timer was still active after C's rendering). When C will
arrive, we would still have A's animation in
currentViewAnimation and thus we will properly replace
B with C in viewToDisplayAfterAnimation.