Integrate android java code with C++ code
Create a libserio-android that stands between libcore
and actual android Java runtime. libserio-android is
a glue code between JNI and Java part of the application
and libcore. It provides API for java part to send events
to libcore and allows libcore to send events back to Java.
Make "core" target a STATIC library, since it is not used
directly in android build anymore.