Drastically improve reliability of animations
Re-implement animations with CSS animations instead of
transitions to reliably execute transition-in animation
without any javascript involvement: previously setTimeout(0)
that was supposed to remove animation CSS classes from the
element and start those animations could execute before
the element was actually attached to DOM (in Chrome), which
led to transition-in animations not being played at all
from time to time. With CSS animations the whole
transition-in animation is handled exclusively on CSS side,
which guarantees it's reliable execution every time.
Replace the reminder of setTimeout() calls in animation
handling with onanimationend callbacks, to further improve
reliability (and as a nice side effect, animation handling
JS does not need to calculate animation duration anymore).