Pass view rendering tasks from C++ to js
Also stub window.core with a dummy core in case such object
is not initialized: if it is not initialized - it means
UI is running in a standalone browser. In such case we
don't need to handle UI events, but at least need to provide
a stub, so that UI does not fail.