Implement fatal error handling on C++ side
Whenever an exception (all programmer errors in Serio are
exceptions) occurs:
- inside main loop
- when sending event to core
- when sending back event to core
display_fatal_error() gets called, that displays dialog with
the message of the error that has occurred. That function
is completely autonomous and does not rely on any other
module (which might be malfunctioning at this point).
It creates and sends error dialog render event completely
on it's own. Displayed dialog does not have buttons and
backEvent, which means the only thing the user could do
is restart the application.