Refine data structure naming
Core sends "render tasks" (not events) to UI, so that UI
can render views, described in those tasks.
UI sends "tasks" (not events) back to core for processing.
Every render task has "viewId" (not just "view"), which is
an ID of a view, that needs to be rendered.
Every task has "taskId" (not "event"), which is an ID of a
task, that needs to be executed.