Implement initial version of "first-boot" use case
When user starts application for the first time (and
"source-url" is not yet specified in database) - ask
the user to specify crawler config URL. Try to fetch the
specified config. If success - store the URL in the database.
If failure - display error dialog.
For now config consumes all incoming http responses. This
will be addressed in later commits.
Rename all view_id values to end with *_view to not clash
with other names.
Rename all task_id values to end with *_task to not clash
with other names.
Add "loading" sub-struct to ui_data, so that we will be
actually able to render loading_screen_view.
Make ui not render same ui_data twice: when new ui_data is
created (by some module to initiate rendering of a new view)
rendered flag is set to false. This will tell ui to create
render tasks for that data. After tasks are created - ui
resets rendered flag to true. Next time when ui receives
the same ui_data unchanged - it won't try to re-render it.
Make http client at least somewhat handle case, when
specified URL is malformed and an http request can't be send
at all.