Make fetch_crawler_config() not consume ALL http responses
Make fetch_crawler_config() send http requests with ID of
the original "download_config_task", so that when responses
arrive, it is able to find the exact response. Store
"active_task" in core, which is the task, that is currently
being processed by the core across multiple
execute_core_task() calls.