Download config in when app starts and URL is already specified
Make config display title screen on it's own for now. I'm
not sure if it's the right place to do that though.
Slightly refactor config.cpp.
Use same cache-ttl both when we download an actual existing
config and when we try to download a new config, specified
by the user: in the second case if download fails - we won't
cache it anyway and if download succeeds - we will cache
it right away and won't need to re-request it after next
restart the same day.
Make http try to return outdated cached response even in
case of StatusCodeInvalid: it could be that user does not
have an internet connection - in such case it's okay to
grab config from cache.