Don't attempt to get expired config from cache on init_task
User specifies config URL for the first time and it works.
Next time user opens the app up, we download the config
and it is not a valid JSON anymore (config maintainer
has screwed up). In such case if we attempt to get previous
response from cache - we get the new broken response,
since http system has already cached it (since
it is valid from http standpoint).
Technically we can store multiple cache entries per key
(in their chronological order), but that will explode
the database size. We would then have to implement
automatic cache garbage collection (while right now we
might be okay without it).
Since this is only needed in this corner case, which is
pretty particular (not only the config maintainer needs
to screw up but also the user needs to manually clear
the cache while the config is broken).
Might implement later, but definitely not now (to keep
things simple).